WIP: add event collection for init-local
init_local has little for benchmarking (events collected for analyze). This adds
some support, however it looks like refactoring the `blame` and `show` code to
support nested calls to the event collector. For now I think the time diffs are
accurate, but the stages get broken into multiple boot records, so something isn't
quite right. For now I'm not going to work on this any more.